Codebase reorganization.
