+++ ios/BTE/DebugAction.swift
@@ -1,0 +1,12 @@
+@objc enum DebugAction: Int {
+  case fetchDiagnosisKeys,
+  detectExposuresNow,
+  simulateExposureDetectionError,
+  simulateExposure,
+  simulatePositiveDiagnosis,
+  disableExposureNotifications,
+  resetExposureDetectionError,
+  resetUserENState,
+  getAndPostDiagnosisKeys,
+  getExposureConfiguration
+}
