+++ ios/BTE/ExposureManager.swift
@@ -261,16 +261,3 @@ final class ExposureManager: NSObject {
   }
   
 }
-
-@objc enum DebugAction: Int {
-  case fetchDiagnosisKeys,
-  detectExposuresNow,
-  simulateExposureDetectionError,
-  simulateExposure,
-  simulatePositiveDiagnosis,
-  disableExposureNotifications,
-  resetExposureDetectionError,
-  resetUserENState,
-  getAndPostDiagnosisKeys,
-  getExposureConfiguration
-}
