+++ ios/BTE/Extensions/String+Extensions.swift
@@ -1,6 +1,7 @@
 import Foundation
 
 extension String {
+  static let `default` = ""
   static let notAuthorized = "notAuthorized"
   static let authorized = "authorized"
 }
